I did a mental analysis of the process that draws a line, this process invokes the process that draws a point on the screen, which when finished puts the GLCD task to idle and it is not correct, since it must draw the following points that make up the line.
So, drawing a point is a subprocess of a larger process, and when it finishes, it must return to a state of the task to complete the larger process. That is the reason for the stateToReturn variable.

Now, it is possible that drawing a line is a subprocess of another higher process, for example drawing a rectangle. If we get to this part, I will have to make a modification similar to what I am describing.

Also when the larger or higher process finishes, it is it that must put the GLD task to sleep, for now it will be the process that draws the line that finishes everything.